feat: :sparkles: add DenunciaRequest class
Create a new file to implement the DenunciaRequest class. This class represents the request payload for creating a denuncia. It includes properties such as `titulo`, `descricao`, `latitude`, `longitude`, and `denunciante`, which are used to provide the denuncia's title, description, location coordinates, and the associated denunciante information. The `DenuncianteRequest` class is used as a nested object within this class.