feat: :sparkles: add MapQuestMapper class
Create a new file to implement the MapQuestMapper class. This class is responsible for mapping the response from the MapQuest API to the Endereco (address) object. The `mapResponseToResult` method takes the API response data and extracts relevant information such as postal code, street, neighborhood, city, state, and country, and returns an Endereco object with these mapped properties.