feat: :sparkles: add MapQuestService class
Create a new file to implement the MapQuestService class. This class is responsible for interacting with the MapQuest API to perform reverse geocoding. It uses the provided latitude and longitude coordinates to build a request URL and retrieve the response from the API. The ConfigService is used to retrieve the MapQuest API URL and key. The HttpService is utilized to make the HTTP request, and the MapQuestMapper class is injected to map the API response to the Endereco object. The `getReverseGeocoding` method performs the reverse geocoding operation and returns the mapped Endereco object.