feat: :sparkles: add DenunciaController class
Create a new file to implement the DenunciaController class. This class serves as a NestJS controller for handling HTTP requests related to denuncias. It includes a `create` method decorated with the `@Post` decorator to handle POST requests to the '/denuncias' endpoint. The `@HttpCode(201)` decorator sets the HTTP response status code to 201 (Created). The method receives a `denunciaRequest` object in the request body and delegates the creation of the denuncia to the DenunciaService.