feat: :sparkles: add DenunciaModule class
Create a new file to implement the DenunciaModule class. This class serves as a NestJS module that encapsulates the dependencies and configuration for the Denuncia feature. It imports the PrismaModule and MapQuestModule to provide the necessary services and configurations for working with the database and MapQuest API. The DenunciaController and DenunciaService are registered as controllers and providers, respectively, within the module.