refactor: ♻️ refactor AppModule
Refactor the AppModule by adding new imports and removing unnecessary imports. In the old commit, the AppModule only imported the AppController and AppService. In the new commit, the AppModule imports additional modules such as DenunciaModule, PrismaModule, MapQuestModule, ConfigModule, and HttpModule. These new imports provide the necessary dependencies and configurations for the application. The controllers and providers arrays have been emptied as there are no specific controllers or providers registered directly in the AppModule.