refactor: ♻️ refactor MapQuestController
Refactor the parameter names in the getAddressFromCoordinates method of MapQuestController for better clarity and consistency. The latitude and longitude query parameters are renamed to lat and long respectively. Additionally, update the variable name 'endereco' to 'address' to improve code readability and maintainability. These changes enhance the overall code quality while preserving the functionality.